Add files via upload
We do this conversion by creating a firstPos, lastPos, and followPos table. Once we have a followPos table, we generate a DFA from that. We then match the input string with the regular expression and echo the characters that do not match. The characters that follow a '$' match the regex, and those that follow a '@' do not.